ILE: Have either LowPower or ND Filters
ND Filters is supposed to be an improvement over LowPower mode.
Make sure to only display the LowPower property if ND Filters is not
supported.

We assume ND Filters is supported if the new power management
interface is available and there is more than 1 low power level.
